requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
bagit==1.7.0
bagit-profile==1.3.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Flask==1.0.3
fonttools==3.43.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.18
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lxml==4.3.4
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.4
ocrd==1.0.0b9
ocrd-modelfactory==1.0.0b9
ocrd-models==1.0.0b9
ocrd-tesserocr==0.2.2
ocrd-utils==1.0.0b9
ocrd-validators==1.0.0b9
opencv-python-headless==4.1.0.25
packaging==19.0
parsel==1.5.1
Pillow==6.0.0
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.3
PyYAML==5.1.1
queuelib==1.5.0
requests==2.22.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.1
tesserocr==2.4.0
Twisted==19.2.1
urllib3==1.25.3
w3lib==1.20.0
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
zope.interface==4.6.0
